Optimize images pre-download for multiple impl of core control
